chore: rename test.js to debugRoutes.js
