+++ bcsc-service/src/routes/debugRoutes.js
